+++ packages/digit_data_model/lib/data/local_store/sql_store/sql_store.g.dart
@@ -17421,7 +17421,8 @@ class $IdentifierTable extends Identifier
   }
 
   @override
-  Set<GeneratedColumn> get $primaryKey => {auditCreatedBy, clientReferenceId};
+  Set<GeneratedColumn> get $primaryKey =>
+      {auditCreatedBy, clientReferenceId, identifierType, identifierId};
   @override
   IdentifierData map(Map<String, dynamic> data, {String? tablePrefix}) {
     final effectivePrefix = tablePrefix != null ? '$tablePrefix.' : '';