+++ packages/digit_data_model/CHANGELOG.md
@@ -23,6 +23,13 @@
 * Updated additionalDetails field to service definition attributes model
 * Changed the object type of additionalDetails in Service model from String to Map
 
+## 1.0.4-dev.5-uniqueId
+* Added identifierType enum entity
+* Added identifierType and id as primary key to identifier table
+
+* Updated oplog transactions to synchronous - fix for write transaction lock
+* Same change in 1.0.3-hotfix, 1.0.4-dev.3
+
 ## 1.0.3+2
 
 * Updated beneficiary type target model and table to string